debugtrace-env/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.10.0
pytest==8.1.1
requests==2.31.0
google-generativeai==0.3.0
